requirements-integration.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements-integration.in
#
aiohappyeyeballs==2.4.4
    # via aiohttp
aiohttp==3.10.11
    # via -r requirements-integration.in
aiosignal==1.3.1
    # via aiohttp
anyio==4.5.2
    # via httpx
asttokens==3.0.0
    # via stack-data
attrs==24.2.0
    # via
    #   aiohttp
    #   jsonschema
backcall==0.2.0
    # via ipython
bcrypt==4.2.1
    # via paramiko
cachetools==5.5.0
    # via google-auth
certifi==2024.8.30
    # via
    #   httpcore
    #   httpx
    #   kfp-server-api
    #   kubernetes
    #   requests
cffi==1.17.1
    # via
    #   cryptography
    #   pynacl
charmed-kubeflow-chisme==0.4.11
    # via -r requirements-integration.in
charset-normalizer==3.4.0
    # via requests
click==8.1.7
    # via kfp
cryptography==44.0.0
    # via paramiko
decorator==5.1.1
    # via
    #   ipdb
    #   ipython
deepdiff==6.2.1
    # via charmed-kubeflow-chisme
deprecated==1.2.18
    # via opentelemetry-api
docstring-parser==0.16
    # via kfp
executing==2.1.0
    # via stack-data
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
google-api-core==2.23.0
    # via
    #   google-cloud-core
    #   google-cloud-storage
    #   kfp
google-auth==2.36.0
    # via
    #   google-api-core
    #   google-cloud-core
    #   google-cloud-storage
    #   kfp
    #   kubernetes
google-cloud-core==2.4.1
    # via google-cloud-storage
google-cloud-storage==2.19.0
    # via kfp
google-crc32c==1.5.0
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-resumable-media==2.7.2
    # via google-cloud-storage
googleapis-common-protos==1.66.0
    # via google-api-core
h11==0.14.0
    # via httpcore
httpcore==1.0.7
    # via httpx
httpx==0.27.2
    # via
    #   charmed-kubeflow-chisme
    #   lightkube
hvac==2.3.0
    # via juju
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
importlib-metadata==8.5.0
    # via
    #   opentelemetry-api
    #   ops
iniconfig==2.0.0
    # via pytest
ipdb==0.13.13
    # via pytest-operator
ipython==8.12.3
    # via ipdb
jedi==0.19.2
    # via ipython
jinja2==3.1.4
    # via
    #   charmed-kubeflow-chisme
    #   pytest-operator
jsonschema==4.17.3
    # via
    #   -r requirements-integration.in
    #   serialized-data-interface
juju==3.6.0.0
    # via
    #   -r requirements-integration.in
    #   charmed-kubeflow-chisme
    #   pytest-operator
kfp==2.9.0
    # via -r requirements-integration.in
kfp-pipeline-spec==0.4.0
    # via kfp
kfp-server-api==2.3.0
    # via kfp
kubernetes==30.1.0
    # via
    #   juju
    #   kfp
lightkube==0.15.6
    # via
    #   -r requirements-integration.in
    #   charmed-kubeflow-chisme
lightkube-models==1.31.1.8
    # via lightkube
macaroonbakery==1.3.4
    # via juju
markupsafe==2.1.5
    # via jinja2
matplotlib-inline==0.1.7
    # via ipython
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via typing-inspect
oauthlib==3.2.2
    # via
    #   kubernetes
    #   requests-oauthlib
opentelemetry-api==1.32.1
    # via ops
ops==2.21.1
    # via
    #   charmed-kubeflow-chisme
    #   serialized-data-interface
ordered-set==4.1.0
    # via deepdiff
packaging==24.2
    # via
    #   juju
    #   pytest
paramiko==3.5.0
    # via juju
parso==0.8.4
    # via jedi
pexpect==4.9.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pluggy==1.5.0
    # via pytest
prompt-toolkit==3.0.48
    # via ipython
propcache==0.2.0
    # via yarl
proto-plus==1.25.0
    # via google-api-core
protobuf==4.25.5
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   kfp
    #   kfp-pipeline-spec
    #   macaroonbakery
    #   proto-plus
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.3
    # via stack-data
pyasn1==0.6.1
    # via
    #   juju
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.1
    # via google-auth
pycparser==2.22
    # via cffi
pygments==2.18.0
    # via ipython
pymacaroons==0.13.0
    # via macaroonbakery
pynacl==1.5.0
    # via
    #   macaroonbakery
    #   paramiko
    #   pymacaroons
pyrfc3339==1.1
    # via
    #   juju
    #   macaroonbakery
pyrsistent==0.20.0
    # via jsonschema
pytest==8.3.4
    # via
    #   -r requirements-integration.in
    #   pytest-asyncio
    #   pytest-operator
pytest-asyncio==0.21.2
    # via pytest-operator
pytest-operator==0.38.0
    # via -r requirements-integration.in
python-dateutil==2.9.0.post0
    # via
    #   kfp-server-api
    #   kubernetes
pytz==2024.2
    # via pyrfc3339
pyyaml==6.0.2
    # via
    #   -r requirements-integration.in
    #   juju
    #   kfp
    #   kubernetes
    #   lightkube
    #   ops
    #   pytest-operator
    #   serialized-data-interface
requests==2.32.3
    # via
    #   google-api-core
    #   google-cloud-storage
    #   hvac
    #   kubernetes
    #   macaroonbakery
    #   requests-oauthlib
    #   requests-toolbelt
    #   serialized-data-interface
requests-oauthlib==2.0.0
    # via kubernetes
requests-toolbelt==0.10.1
    # via kfp
rsa==4.9
    # via google-auth
ruamel-yaml==0.18.10
    # via charmed-kubeflow-chisme
ruamel-yaml-clib==0.2.8
    # via ruamel-yaml
serialized-data-interface==0.7.0
    # via charmed-kubeflow-chisme
sh==2.1.0
    # via -r requirements-integration.in
six==1.17.0
    # via
    #   kfp-server-api
    #   kubernetes
    #   macaroonbakery
    #   pymacaroons
    #   python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
stack-data==0.6.3
    # via ipython
tabulate==0.9.0
    # via kfp
tenacity==9.0.0
    # via
    #   -r requirements-integration.in
    #   charmed-kubeflow-chisme
toposort==1.10
    # via juju
traitlets==5.14.3
    # via
    #   ipython
    #   matplotlib-inline
typing-extensions==4.12.2
    # via
    #   juju
    #   typing-inspect
typing-inspect==0.9.0
    # via juju
urllib3==1.26.20
    # via
    #   kfp
    #   kfp-server-api
    #   kubernetes
    #   requests
wcwidth==0.2.13
    # via prompt-toolkit
websocket-client==1.8.0
    # via
    #   kubernetes
    #   ops
websockets==13.1
    # via juju
wrapt==1.17.2
    # via deprecated
yarl==1.15.2
    # via aiohttp
zipp==3.20.2
    # via importlib-metadata
